new price tracker smooth.py
I scrapped the entire code and rewrote it to be much smoother and faster. No longer has errors(unless the websites themselves are down, although I don't have control over that) and as a final touch i remade the GUI using the PyQt library for a more modern look. In the future I will add more websites for access and a feature that provides links for all of the products at the end of the program